companion/requirements.txt
# Core service
Flask>=2.3
requests>=2.31
python-dotenv>=1.0

# ML stack
scikit-learn>=1.3
joblib>=1.3
numpy>=1.24
pandas>=2.0

# Optional RAG (comment out if not used)
# sentence-transformers>=2.2
# faiss-cpu>=1.7

# Optional Gemini LLM (only needed if GEMINI_ENABLED=true)
google-generativeai>=0.8.0
